Remove magical names for anyref items
This commit removes all the magical names flying around in the JS
generation for managing anyref items and anyref tables. Instead IDs are
set during the anyref pass and these are propagated to the end. This
also deletes references to GC'ing code in the anyref module now that
they've been verified to no longer be needed.