cli-support: Remove Node.js specific passStringToWasm
Now that Node has encodeInto it can use the same path as the normal Web
Browser code (that avoids calling into the C++ code). Node 12,13, and 14
have this method, so it is now fairly well supported.

We should check to see if this is actually a speedup for ASCII strings

Signed-off-by: Joe Richey <joerichey@google.com>