Correctly consume tokens when parsing `js_namespace`
`syn` < 1.0.66 would consume the `AnyIdent` anyway while trying to parse
the `ExprArray`.

Now due to a refactoring they did, they (correctly) leave the `ParseBuffer`
untouched if the input is not a valid `ExprArray`, so we should avoid forking
the input buffer and just use the same for both tries, otherwise the unwanted
tokens left in there would make `Punctuated` break soon after.

Fixes #2508