web-sys: update Geolocation webidl files
update the Geolocation IDL files based on

https://www.w3.org/TR/geolocation

The following structures are renamed
* Position -> GeolocationPosition
* PositionError -> GeolocationPositionError
* Coordinates -> GeolocationCoordinates

Browser support according to MDN
* Geolocation
** Chrome  50 (2016-04-13)
** Edge    79 (2020-01-15)
** Firefox 55 (2017-08-08)
* GeolocationPosition
** Chrome  79 (2019-12-10)
** Edge    79 (2020-01-15)
** Firefox 72 (2020-01-07)
* GeolocationPositionError
** Chrome  79 (2019-12-10)
** Edge    79 (2020-01-15)
** Firefox 72 (2020-01-07)
* GeolocationCoordinates
** Chrome  79 (2019-12-10)
** Edge    79 (2020-01-15)
** Firefox 72 (2020-01-07)
* PositionOptions
** Chrome  50 (2016-04-13)
** Edge    79 (2020-01-15)
** Firefox 55 (2017-08-08)

The dates are either for the interface renaming,
or for when the required secure-context for the API is supported.

According to [CanIUse](https://caniuse.com/geolocation) linked in the spec,
the API should also be supported by all major browsers.

related GH issue: #2491