Don't create duplicate closure adapters for the same function signature
Fixes #2830

This was causing things to break with reference types enabled, since each closure adapter had a corresponding wasm export, leading to duplicate wasm exports, which is invalid.

I did this by adding some deduplication for closures (both long-lived and stack).

There was already some deduplication for long-lived closures (which I removed in favour of the shared deduplication), but it considered named externrefs different to unnamed externrefs, which the compiler doesn't. So with the new deduplication, I first replace all named externrefs with unnamed externrefs to make sure they're treated the same.