Speed up the check for whether we need to remake a memory view
Getting the buffer of a `WebAssembly.Memory` seems to have some overhead to it, so instead check if the existing view is 0 bytes long, which happens when they become detached after the memory is grown.

I'm pretty sure that they'll never be zero bytes long otherwise; but even if they are, it's not the end of the world to recreate the view every time in that rare circumstance.