Initialise memory views when wasm is initialised
With the last commit, if `init` was called twice on web/no-modules, the memory views would be stuck pointing to the first copy of memory, since it never actually gets detached; this fixes that by manually resetting them.

This also allows eliminating the null check in the memory view getters, but it has practically no effect on performance.