Don't apply doc comments to `extern` blocks
This was causing lots of 'unused doc comment' warnings while compiling `web-sys`, since doc comments have no effect on `extern` blocks.

I've implemented this by removing any doc comments from the function's attributes and then adding them back to the public function.

I'm not sure how good of an idea it is to forward all attributes to the extern block in the first place, but I'm playing it safe and leaving it as is.